+++ plugin/src/main/kotlin/co/infinum/collar/plugin/tasks/GenerateTask.kt
@@ -2,22 +2,46 @@ package co.infinum.collar.plugin.tasks
 
 import co.infinum.collar.generator.CollarGenerator
 import co.infinum.collar.plugin.CollarExtension
-import co.infinum.collar.plugin.tasks.shared.BaseTask
+import co.infinum.collar.plugin.tasks.shared.BaseSourceTask
 import co.infinum.collar.plugin.validate
+import com.android.builder.model.AndroidProject.FD_GENERATED
+import org.gradle.api.file.FileTree
+import org.gradle.api.tasks.CacheableTask
+import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.InputFiles
+import org.gradle.api.tasks.OutputDirectory
+import org.gradle.api.tasks.PathSensitive
+import org.gradle.api.tasks.PathSensitivity
+import org.gradle.api.tasks.SkipWhenEmpty
 import org.gradle.api.tasks.TaskAction
+import java.io.File
 
-internal open class GenerateTask : BaseTask() {
+@CacheableTask
+internal open class GenerateTask : BaseSourceTask() {
 
     companion object {
         const val GROUP = "collar"
         const val NAME = "generate"
         const val DESCRIPTION = "Generates Kotlin files for screen names, events and user properties."
-
-        private const val TEMPLATE_OUTPUT_PATH = "%s/src/%s/kotlin"
-        private const val TEMPLATE_FILE_PATH = "%s/%s"
     }
 
+    @Suppress("unused") // Required to invalidate the task on version updates.
+    @Input
+    val pluginVersion = CollarExtension.DEFAULT_VERSION
+
+    @get:OutputDirectory
+    var outputDirectory: File = File(
+        project.buildDir,
+        "$FD_GENERATED${File.separatorChar}${CollarExtension.NAME}${File.separatorChar}trackingPlan"
+    )
+
     private val collarGenerator = CollarGenerator()
+
+    @InputFiles
+    @SkipWhenEmpty
+    @PathSensitive(PathSensitivity.ABSOLUTE)
+    override fun getSource(): FileTree =
+        super.getSource()
 
     @TaskAction
     fun doOnRun() {
@@ -26,25 +50,23 @@ internal open class GenerateTask : BaseTask() {
             when {
                 errors.isNotEmpty() -> showError(errors.joinToString(System.lineSeparator()))
                 else -> {
-                    val outputPath = String.format(TEMPLATE_OUTPUT_PATH, project.projectDir, variant)
-                    val filePath = String.format(TEMPLATE_FILE_PATH, project.projectDir, filePath)
-                    generateTrackingPlan(filePath, outputPath, packageName)
+                    generateTrackingPlan(packageName)
                 }
             }
         }
     }
 
     @Suppress("TooGenericExceptionCaught")
-    private fun generateTrackingPlan(filePath: String, outputPath: String, packageName: String) {
+    private fun generateTrackingPlan(packageName: String) {
         try {
-            println("Tracking plan file path: $filePath")
-            if (collarGenerator.generate(filePath, outputPath, packageName)) {
-                println("Tracking plan classes generated on path: $outputPath")
+            println("Tracking plan file path: ${source.first().absolutePath}")
+            if (collarGenerator.generate(source.first().absolutePath, outputDirectory.absolutePath, packageName)) {
+                println("Tracking plan classes generated on path: ${outputDirectory.absolutePath}")
             } else {
                 showError("Task generate failed")
             }
-        } catch (e: Exception) {
-            e.printStackTrace()
+        } catch (throwable: Throwable) {
+            throwable.printStackTrace()
         }
     }
 }